% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkoi-package.R
\docType{package}
\name{tkoi-package}
\alias{tkoi}
\alias{tkoi-package}
\title{tkoi: Transcriptomic Knowledge-Graph Omics Integration}
\description{
Network-aware gene enrichment analysis. Transcriptomic signals are
propagated through a human biological knowledge graph with personalized
PageRank, and a degree-matched permutation null identifies enriched
biological concepts. The main entry point is \code{\link[=run_tkoi]{run_tkoi()}}.
}
\section{Performance}{

The PageRank solver, neighborhood counting, and null sampling are
implemented in C++ and run on multiple threads. See the \code{n_cores}
argument of \code{\link[=run_tkoi]{run_tkoi()}} and the \code{tkoi.n_cores} option.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BaranziniLab/tkoi}
  \item Report bugs at \url{https://github.com/BaranziniLab/tkoi/issues}
}

}
\author{
\strong{Maintainer}: Wanjun Gu \email{wanjun.gu@ucsf.edu} (\href{https://orcid.org/0000-0002-7342-7000}{ORCID})

}
\keyword{internal}
